feat: split Docker and Docker Swarm
