+++ pkg/redactor/redactor_config_test.go
@@ -39,7 +39,7 @@ import (
 	"github.com/traefik/traefik/v2/pkg/types"
 )
 
-var updateExpected = flag.Bool("update_expected", true, "Update expected files in fixtures")
+var updateExpected = flag.Bool("update_expected", false, "Update expected files in fixtures")
 
 var fullDynConf *dynamic.Configuration
 
@@ -591,22 +591,28 @@ func TestDo_staticConfiguration(t *testing.T) {
 	}
 
 	config.Providers.Docker = &docker.Provider{
-		Constraints: `Label("foo", "bar")`,
-		Watch:       true,
-		Endpoint:    "MyEndPoint",
-		DefaultRule: "PathPrefix(`/`)",
-		TLS: &types.ClientTLS{
-			CA:                 "myCa",
-			Cert:               "mycert.pem",
-			Key:                "mycert.key",
-			InsecureSkipVerify: true,
+		Shared: docker.Shared{
+			ExposedByDefault:   true,
+			Constraints:        `Label("foo", "bar")`,
+			AllowEmptyServices: true,
+			Network:            "MyNetwork",
+			UseBindPortIP:      true,
 		},
-		ExposedByDefault:        true,
-		UseBindPortIP:           true,
-		SwarmMode:               true,
-		Network:                 "MyNetwork",
-		SwarmModeRefreshSeconds: 42,
-		HTTPClientTimeout:       42,
+		ClientConfig: docker.ClientConfig{
+			Endpoint: "MyEndPoint", TLS: &types.ClientTLS{
+				CA:                 "myCa",
+				Cert:               "mycert.pem",
+				Key:                "mycert.key",
+				InsecureSkipVerify: true,
+			},
+			HTTPClientTimeout: 42,
+		},
+
+		Watch:       true,
+		DefaultRule: "PathPrefix(`/`)",
+
+		// SwarmMode:               true,
+		// SwarmModeRefreshSeconds: 42,
 	}
 
 	config.Providers.KubernetesIngress = &ingress.Provider{
